refactor(api): serve API reference at docs path
