+++ web/src/pages/api/docs.ts
